004 ServerCode/DB/requirements.txt
alembic==1.16.1
altgraph==0.17.4
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
greenlet==3.2.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
onnxruntime==1.21.1
opencv-python-headless==4.11.0.86
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.30.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyMatting==1.1.13
pyreadline3==3.5.4
pywin32-ctypes==0.2.3
referencing==0.36.2
rembg==2.0.65
requests==2.32.3
rpds-py==0.24.0
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.4.0
SQLAlchemy==2.0.41
sympy==1.14.0
tifffile==2025.5.10
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
